+++ actions/products/delete-product.ts
@@ -1,10 +1,17 @@
 "use server"
 
+import { auth } from "@/lib/auth/auth"
 import prisma from "@/lib/db/db"
 import { revalidatePath } from "next/cache"
 import deleteImage from "../cloudinary/delete-image"
 
 export async function deleteProduct({ id }: { id: string }) {
+  const session = await auth()
+
+  if (session?.user.role !== "ADMIN") {
+    return { error: "No autorizado." }
+  }
+
   try {
     const product = await prisma.product.delete({
       where: { id },
