+++ actions/promotions/delete-promotion.ts
@@ -1,9 +1,16 @@
 "use server"
 
+import { auth } from "@/lib/auth/auth"
 import prisma from "@/lib/db/db"
 import { revalidatePath } from "next/cache"
 
 export async function deletePromotion({ id }: { id: string }) {
+  const session = await auth()
+
+  if (session?.user.role !== "ADMIN") {
+    return { error: "No autorizado." }
+  }
+
   try {
     const promotion = await prisma.promotion.delete({
       where: { id },
