+++ actions/recipes/delete-recipe.ts
@@ -1,9 +1,16 @@
 "use server"
 
+import { auth } from "@/lib/auth/auth"
 import prisma from "@/lib/db/db"
 import { revalidatePath } from "next/cache"
 
 export async function deleteRecipe({ id }: { id: string }) {
+  const session = await auth()
+
+  if (session?.user.role !== "ADMIN") {
+    return { error: "No autorizado." }
+  }
+
   try {
     const recipe = await prisma.recipe.delete({
       where: { id },
